Move job state model out of mvc/history to utils
