+++ client/src/components/JobStates/wait.js
@@ -1,5 +1,5 @@
 import { getAppRoot } from "onload/loadConfig";
-import JOB_STATES_MODEL from "mvc/history/job-states-model";
+import JOB_STATES_MODEL from "utils/job-states-model";
 import axios from "axios";
 
 export function waitOnJob(jobId, onStateUpdate = null, interval = 1000) {
