+++ client/src/components/admin/Jobs.vue
@@ -106,7 +106,7 @@ import { getAppRoot } from "onload/loadConfig";
 import axios from "axios";
 import JobsTable from "components/admin/JobsTable";
 import JobLock from "./JobLock";
-import JOB_STATES_MODEL from "mvc/history/job-states-model";
+import JOB_STATES_MODEL from "utils/job-states-model";
 import { commonJobFields } from "./JobFields";
 import { errorMessageAsString } from "utils/simple-error";
 import { jobsProvider } from "components/providers/JobProvider";
